Move blk_jacobi into its own file.
In the future, we'll have more smoothers, so we might as well start organizing
them now.
