+++ py_simplemg/smoothers.py
@@ -1,0 +1,24 @@
+
+import numpy as np
+
+def blk_jacobi(A, x, b, redblack=True):
+  """ Performs one block Jacobi iteration.  Red-black ordering can be toggled
+      on/off. """
+  tmpdiag = np.diag(A)
+  tmpdiag = 1./tmpdiag
+  if redblack:
+    tmpdiag[1::2] = 0.
+  tmpdiag = np.diag(tmpdiag)
+
+  x = x+np.dot(tmpdiag, b-np.dot(A, x))
+  if not redblack:
+    return x
+
+  tmpdiag2 = np.diag(A)
+  tmpdiag2 = 1./tmpdiag2
+  tmpdiag2[0::2] = 0.
+  tmpdiag2 = np.diag(tmpdiag2)
+
+  x = x+np.dot(tmpdiag, b-np.dot(A, x))
+
+  return x
